clio/backend/requirements.txt
Django>=5.1,<5.3
djangorestframework>=3.15
drf-spectacular>=0.27
django-filter>=24.0
django-redis>=5.4
django-cors-headers>=4.3
redis>=5.0
psycopg[binary]>=3.1
gunicorn>=22.0
uvicorn>=0.30
httpx>=0.27
cryptography>=42.0
PyJWT>=2.8
passlib>=1.7
boto3>=1.34
